requirements.txt
# AI角色扮演聊天网站依赖包

# 基础Web框架
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Werkzeug==2.3.7

# HTTP请求和环境配置
requests==2.31.0
python-dotenv==1.0.0

# 语音识别和合成
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.11

# 音频处理
pydub==0.25.1
numpy==1.24.3

# 实时通信
websockets==11.0.3

# AI和数据处理
openai==1.3.0
httpx==0.24.1

# 可选依赖（用于更好的音频处理）
# webrtcvad==2.0.10  # 语音活动检测
# scipy==1.11.1     # 高级音频分析